requirement.txt
openai
PyGithub
python-dotenv
discord.py
flask
